refactored FilterSummary to just be FilteringFirstPass after learning
